- get multiple slots from StateDatabase - extract AbstractStateDatabaseTest and inherit MemoryStateDatabaseTest and FileSystemStateDatabaseTest from it - fix MemoryStateDatabase so those tests pass
