+++ celos-server/src/main/java/com/collective/celos/StateDatabase.java
@@ -15,9 +15,11 @@
  */
 package com.collective.celos;
 
+import com.google.common.collect.Maps;
+
 import java.io.IOException;
 import java.util.Collection;
-import java.util.List;
+import java.util.Map;
 import java.util.SortedSet;
 
 /**
@@ -26,16 +28,25 @@ import java.util.SortedSet;
 public interface StateDatabase {
 
     /**
-     * Returns the state of the slots, specified by start time (inclusive) and end time (exclusive)
-     * or empty List if not found.
+     * Returns the state of the given slots, or empty Map if not found.
      */
-    public List<SlotState> getSlotStates(WorkflowID id, ScheduledTime start, ScheduledTime end) throws Exception;
+    public Map<SlotID, SlotState> getSlotStates(WorkflowID id, ScheduledTime start, ScheduledTime end) throws Exception;
 
 
     /**
-     * Returns the state of the given slots, or empty List if not found.
+     * Returns the state of the given slots, or empty Map if not found.
      */
-    public List<SlotState> getSlotStates(WorkflowID id, Collection<ScheduledTime> times) throws Exception;
+    public default Map<SlotID, SlotState> getSlotStates(WorkflowID id, Collection<ScheduledTime> times) throws Exception {
+        Map<SlotID, SlotState> slotStates = Maps.newHashMap();
+        for (ScheduledTime time : times) {
+            SlotID slotID = new SlotID(id, time);
+            SlotState slotState = getSlotState(slotID);
+            if (slotState != null) {
+                slotStates.put(slotID, slotState);
+            }
+        }
+        return slotStates;
+    }
 
     /**
      * Returns the state of the given slot, or null if not found.
@@ -44,7 +55,7 @@ public interface StateDatabase {
 
     /**
      * Updates the state of the given slot.
-     *
+     * <p>
      * If this is a rerun, then markSlotForRerun() must be used in addition.
      */
     public void putSlotState(SlotState state) throws Exception;
