+++ celos-server/src/test/java/com/collective/celos/MemoryStateDatabase.java
@@ -15,10 +15,11 @@
  */
 package com.collective.celos;
 
-import com.google.common.collect.Lists;
+import com.google.common.collect.Maps;
 import org.joda.time.DateTime;
 
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Simple implementation of StateDatabase that stores everything in a Map.
@@ -26,28 +27,29 @@ import java.util.*;
 public class MemoryStateDatabase implements StateDatabase {
 
     protected final Map<SlotID, SlotState> map = new HashMap<SlotID, SlotState>();
-    protected final Set<SlotID> rerun = new HashSet<SlotID>();
+    protected final Map<SlotID, ScheduledTime> rerun = new ConcurrentHashMap<>();
     protected final Set<WorkflowID> pausedWorkflows = new HashSet<>();
 
     @Override
-    public List<SlotState> getSlotStates(WorkflowID id, ScheduledTime start, ScheduledTime end) throws Exception {
-        List<SlotState> slotStates = Lists.newArrayList();
-        for (Map.Entry<SlotID, SlotState> entry: map.entrySet()) {
+    public Map<SlotID, SlotState> getSlotStates(WorkflowID id, ScheduledTime start, ScheduledTime end) throws Exception {
+        Map<SlotID, SlotState> slotStates = Maps.newHashMap();
+        for (Map.Entry<SlotID, SlotState> entry : map.entrySet()) {
             DateTime dateTime = entry.getKey().getScheduledTime().getDateTime();
             if (!dateTime.isBefore(start.getDateTime()) && dateTime.isBefore(end.getDateTime())) {
-                slotStates.add(entry.getValue());
+                slotStates.put(entry.getKey(), entry.getValue());
             }
         }
         return slotStates;
     }
 
     @Override
-    public List<SlotState> getSlotStates(WorkflowID id, Collection<ScheduledTime> times) throws Exception {
-        List<SlotState> slotStates = Lists.newArrayList();
+    public Map<SlotID, SlotState> getSlotStates(WorkflowID id, Collection<ScheduledTime> times) throws Exception {
+        Map<SlotID, SlotState> slotStates = Maps.newHashMap();
         for (ScheduledTime time : times) {
-            SlotState slotState = getSlotState(new SlotID(id, time));
+            SlotID slotID = new SlotID(id, time);
+            SlotState slotState = getSlotState(slotID);
             if (slotState != null) {
-                slotStates.add(slotState);
+                slotStates.put(slotID, slotState);
             }
         }
         return slotStates;
@@ -70,23 +72,27 @@ public class MemoryStateDatabase implements StateDatabase {
     @Override
     public void markSlotForRerun(SlotID slot, ScheduledTime now) throws Exception {
         // Doesn't implement GC for rerun
-        rerun.add(slot);
-    }
-
-    @Override
-    public SortedSet<ScheduledTime> getTimesMarkedForRerun(WorkflowID workflowID, ScheduledTime now) throws Exception {
-        SortedSet<ScheduledTime> res = new TreeSet<>();
-        for (SlotID slot : rerun) {
-            if (slot.getWorkflowID().equals(workflowID)) {
-                res.add(slot.getScheduledTime());
-            }
-        }
-        return res;
+        rerun.put(slot, now);
     }
 
     @Override
     public boolean isPaused(WorkflowID workflowID) {
         return pausedWorkflows.contains(workflowID);
+    }
+
+    @Override
+    public SortedSet<ScheduledTime> getTimesMarkedForRerun(WorkflowID workflowID, ScheduledTime now) throws Exception {
+        SortedSet<ScheduledTime> res = new TreeSet<>();
+        for (Map.Entry<SlotID, ScheduledTime> entry : rerun.entrySet()) {
+            if (entry.getKey().getWorkflowID().equals(workflowID)) {
+                res.add(entry.getKey().getScheduledTime());
+                RerunState rerunState = new RerunState(entry.getValue());
+                if (rerunState.isExpired(now)) {
+                    rerun.remove(entry.getKey());
+                }
+            }
+        }
+        return res;
     }
 
     @Override
